refactor: Rename reactions.js to bot.js
