+++ src/bot-core/bot.js
